Support multiple servers in one Sensu check
We run multiple Mesos and Marathon masters, and dozens of slaves.  When
any go down we want an alert.  Currently the Mesos check supports a
single server only.  This commit adds support for multiple servers,
combining all the error messages into a single result string.  It also
adds support for setting a port number.

I tested this locally with different options, but it would be nice to
have unit tests for the various options.  That would be quite a big
change given the current lack of tests, and the need for mock servers,
so I'm punting them to another commit.