Use correct parameter for retrieving running tasks
There is a typo in the url used for the tasks API call. The state
parameter doesn't exist and is replaced by this commit with status,
which is the correct parameter name [0].

Also fixed: A dependency pulls json > 2.0 which depends on ruby > 2.0.
This breaks the 1.9.3 build, so an explicit requirement of json < 2.0
is added.

[0]: https://mesosphere.github.io/marathon/docs/rest-api.html#get-v2-tasks